refactor: migrate script files from CJS to TypeScript ESM
- Convert scripts/compare-results.cjs to scripts/compare-results.ts
- Convert scripts/test-integration.cjs to scripts/test-integration.ts
- Add proper ES module imports and exports
- Use fileURLToPath for __dirname equivalent in ESM
- Remove old .cjs script files to maintain consistency
- This enables modern ES module syntax in all script files
